stocks_products/requirements.txt
django
djangorestframework
django-filter
